Removed unnecessary print line
Forgot to add descriptions to the previous commits. They included a function for updating the power of the AGV's and code to make them charge at the charging nodes when their power is below the threshold. Also, AGV's that have 0 power will be moved to a charging node after 50 time steps.